requirements.txt
django
psycopg
psycopg2-binary
djangorestframework
python-dotenv

